Django Class 01/my_company_project/requirements.txt
djangorestframework
faker